Remove fs dependency
